requirements.txt
langchain
langchain-community
chromadb
sentence-transformers
groq
gradio
requests
python-dotenv
beautifulsoup4
lxml
